Revised Vagrantfile and site.pp
I modified the Vagrantfile to include the boxes I've been testing.  I also scaled down the number of test VMs to two and increased their RAM to 512mb in order to speed things up.  This can be changed back once testing is done.

In site.pp I STATICALLY set aserver.puppetlabs.lan because I found that EVERY VM, regardless of name, was using that node's definition.  Nan checked it out with me yesterday - it was very weird. This should be reset once I test it further.